use 32 bit indices for everything

+++ Source/Engine/Systems/Renderer/Core/Meshes/MeshBufferData.cpp
@@ -6,7 +6,7 @@
 namespace Engine
 {
 
-	void MeshBufferData::GenerateBuffersAndAABB(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices)
+	void MeshBufferData::GenerateBuffersAndAABB(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices)
 	{
 		indexCount = static_cast<uint32_t>(indices.size());
